kafka  producers  added

+++ ingestion-service/src/main/java/com/geopulse/ingestion/service/LocationIngestionService.java
@@ -87,7 +87,7 @@ public class LocationIngestionService {
                 request.timestamp(),
                 request.speed(),
                 request.heading(),
-                request.accuracy().doubleValue(),
+                request.accuracy(),
                 storageCell,
                 partitionCell
         );
